+++ OndselSolver/LDUSpMatParPvPrecise.cpp
@@ -18,17 +18,17 @@ void LDUSpMatParPvPrecise::doPivoting(size_t p)
 	//"Check for singular pivot."
 	//"Do scaling. Do partial pivoting."
 	//| max rowPivot aip mag lookForFirstNonZeroInPivotCol i |
-	int i;	//Use int because of decrement
+	ssize_t i;	//Use ssize_t because of decrement
 	size_t rowPivoti;
 	double aip, mag, max;
 	SpRowDsptr spRowi;
 	rowPositionsOfNonZerosInPivotColumn->clear();
 	auto lookForFirstNonZeroInPivotCol = true;
-	i = (int)m - 1;
+	i = (ssize_t)m - 1;
 	while (lookForFirstNonZeroInPivotCol) {
 		spRowi = matrixA->at(i);
 		if (spRowi->find(p) == spRowi->end()) {
-			if (i <= (int)p) throwSingularMatrixError(""); //Use int because i can be negative
+			if (i <= (ssize_t)p) throwSingularMatrixError(""); //Use ssize_t because i can be negative
 		}
 		else {
 			markowitzPivotColCount = 0;
@@ -41,7 +41,7 @@ void LDUSpMatParPvPrecise::doPivoting(size_t p)
 		}
 		i--;
 	}
-	while (i >= (int)p) { //Use int because i can be negative
+	while (i >= (ssize_t)p) { //Use ssize_t because i can be negative
 		spRowi = matrixA->at(i);
 		if (spRowi->find(p) == spRowi->end()) {
 			aip = std::numeric_limits<double>::min();
